Improve enemy movement, right, to left, and back again.
This is probably a terrible way to do it, but first that comes to mind.
There is a bounding box which is the 'acceptable fleet area', and there is
a fleet controller which just maintains the current heading for the fleet.
Whenever a single ship exits that bounding box, it informs the controller.

This will come back to bite me if I ever try to implement the 'mothership'
enemy from the OG game, as it moves at a higher speed and thus would have
to track its heading separately from the main fleet, ¯\_ (ツ)_/¯.